requirements.txt
langchain
langchain-community
langchain-openai
langgraph
faiss-cpu
sentence-transformers
python-dotenv
tqdm